requirements.txt
pyspark
yfinance
pandas
requests
beautifulsoup4
lxml
python-dotenv
jupyter
pyarrow
certifi